Global Illumination.
Safe-clamp the atmosphere result so as to not cause NaN results.
Extract the tone-mapping functions to a separate file.
Refactor the light pass shader to be more streamlined and concise.
Change the luminance average to be a rotational history buffer.
Change the motion vector to be in the [0-viewport size) screen space.
Sample the light pass luminance correctly in the closest hit shader and apply the distance attenuation and view-cosine factor.
Add an integration pass to convert the radiance probe tiles to 3-band spherical harmonics.
Add a reprojection pass to carry over the T-1 frame's radiance cache.
Refactor the GPU resource clear and copy API to be defer-processed.
Add the per-launch frame counter.
Update the Halton samplers on the fly and use the frame counter to generate the radical inverse numbers.